backend/requirements.txt
asgiref==3.7.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.10.7
idna==3.4
mysql-connector-python==8.0.33
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==10.0.0
platformdirs==3.2.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.4
twilio==7.17.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
